User business logic is completed

+++ src/main/java/nl/avans/drivio/service/UserService.java
@@ -4,10 +4,7 @@ package nl.avans.drivio.service;
 import nl.avans.drivio.model.User;
 import nl.avans.drivio.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.Optional;
@@ -45,6 +42,7 @@ public class UserService {
         User existingUser = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalArgumentException("No user found with given id["+ userId +"]"));
 
+        // set properties to be changed
         existingUser.setFirstName(user.getFirstName());
         existingUser.setLastName(user.getLastName());
         existingUser.setCity(user.getCity());
@@ -60,7 +58,13 @@ public class UserService {
     public void deleteUser(int userId) {
 
         // check whether a user exists in the database. if not, throw exception.
-        userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("No user found with given id["+ userId +"]"));
-        userRepository.deleteById(userId);
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isPresent()) {
+            // delete user with given id
+            userRepository.deleteById(userId);
+        } else {
+            throw new IllegalArgumentException("No user found with given id["+ userId +"]");
+        }
     }
 }
+
